% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{aic}
\alias{aic}
\alias{aic.remstimate}
\title{aic}
\usage{
aic(object, ...)

\method{aic}{remstimate}(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the 'aic' method}
}
\description{
A function that returns the AIC (Akaike's Information Criterion) value in a 'remstimate' object
}
\section{Methods (by class)}{
\itemize{
\item \code{aic(remstimate)}: AIC (Akaike's Information Criterion) value of a 'remstimate' object

}}
