% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{AIC.remstimate}
\alias{AIC.remstimate}
\title{AIC for remstimate objects}
\usage{
\method{AIC}{remstimate}(object, ...)
}
\arguments{
\item{object}{a \code{remstimate} object.}

\item{...}{further arguments passed to \code{\link[stats]{AIC}}.}
}
\value{
AIC value.
}
\description{
Returns the AIC (Akaike's Information Criterion) value of a 'remstimate' object.
}
