% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{aicc}
\alias{aicc}
\alias{aicc.remstimate}
\title{aicc}
\usage{
aicc(object, ...)

\method{aicc}{remstimate}(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the 'aicc' method}
}
\description{
A function that returns the AICC (Akaike's Information Corrected Criterion) value in a 'remstimate' object
}
\section{Methods (by class)}{
\itemize{
\item \code{aicc(remstimate)}: AICC (Akaike's Information Corrected Criterion) value of a 'remstimate' object

}}
