% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{AICC.remstimate}
\alias{AICC.remstimate}
\title{AICC.remstimate()}
\usage{
AICC.remstimate(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object}

\item{...}{further arguments to be passed.}
}
\description{
A function that prints out a summary of the optimization result (codings are based on the structure of summary.lm() and summary.glm() from package 'stats')
}
