% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ao_reh}
\alias{ao_reh}
\title{Actor-Oriented Relational Event History}
\format{
\code{ao_reh} is an \code{reh} object containing the following additional attributes:
\itemize{
   \item{}{\code{attr(ao_reh,"true.pars")}, that is a list of two vectors named \code{"rate_model"} and \code{"choice_model"}, each containing the values of the parameters used in the generation of the event sequence}
   \item{}{\code{attr(ao_reh,"remulate.seed")}, that contains the seed value used in \code{remulate::remulateActor()} for generating the event sequence}
   \item{}{\code{attr(ao_reh,"remulate.reh")}, is a \code{data.frame} with the raw simulated edgelist. The columns of the \code{data.frame} are:
           \itemize{
               \item{}{\code{time}: the timestamp indicating the time at which each event occurred}
               \item{}{\code{actor1}: the actor that generated the relational event}
               \item{}{\code{actor2}: the actor that received the relational event}
           }
    }
}
}
\usage{
data(ao_reh)
}
\description{
The processed \code{reh} object of a randomly generated sequence of relational events with 5 actors and 100 events. The event sequence is generated by following an actor-oriented modeling approach (for more information on the algorithm used for the generation, refer to \code{help(topic = remulateActor, package = "remulate")} or \code{?remulate::remulateActor}).
}
\examples{

# load the data into the workspace
data(ao_reh)

# define linear predictor and claculate stastistcs with \code{remstats} package

## linear predictor for the rate model
rate_model <- ~ 1 + remstats::indegreeSender()

## linear predictror for the choice model
choice_model <- ~ remstats::inertia() + remstats::reciprocity()

## calculate statistics
ao_reh_stats <- remstats::remstats(edgelist = ao_reh, sender_effects = rate_model, 
receiver_effects = choice_model)

## estimate model using method = "MLE"
mle_ao <- remstimate::remstimate(reh = ao_reh, stats = ao_reh_stats, method = "MLE")

## print out a summary of the estimation
summary(mle_ao)

}
\keyword{datasets}
