% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{bic}
\alias{bic}
\alias{bic.remstimate}
\title{bic}
\usage{
bic(object, ...)

\method{bic}{remstimate}(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the 'bic' method}
}
\description{
A function that returns the BIC (Bayesian Information Criterion) value in a 'remstimate' object
}
\section{Methods (by class)}{
\itemize{
\item \code{remstimate}: BIC (Bayesian Information Criterion) value of a 'remstimate' object
}}

