% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{BIC.remstimate}
\alias{BIC.remstimate}
\title{BIC for remstimate objects}
\usage{
\method{BIC}{remstimate}(object, ...)
}
\arguments{
\item{object}{a \code{remstimate} object.}

\item{...}{further arguments passed to \code{\link[stats]{BIC}}.}
}
\value{
BIC value.
}
\description{
Returns the BIC (Bayesian Information Criterion) value of a 'remstimate' object.
}
