% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate_mixrem_doc.R
\name{bic_table}
\alias{bic_table}
\title{Compare BIC across a list of MIXREM fits}
\arguments{
\item{x}{A \code{remstimate_mixrem_list} returned when \code{k} is a
vector in \code{remstimate(..., method = "MIXREM")}.}

\item{...}{Unused.}
}
\value{
A data frame with columns \code{k}, \code{BIC}, and
  \code{delta_BIC}, sorted by ascending BIC.
}
\description{
Compare BIC across a list of MIXREM fits
}
\examples{
\donttest{
library(remstimate)
data(history, package = "remstats")
data(info,    package = "remstats")
colnames(history)[colnames(history) == "setting"] <- "type"

reh   <- remify::remify(edgelist = history[1:100, ], model = "tie",
                         riskset = "active")
stats <- remstats::tomstats(~ inertia(consider_type = FALSE), reh = reh,
                             attr_actors = info)

fits <- remstimate(reh, stats, method = "MIXREM",
                   random = ~ (1 + inertia | dyad), k = 2:5)
bic_table(fits)
}

}
