% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{burninHMC}
\alias{burninHMC}
\title{burninHMC (to check whether this function experiences issues with the definition of int rows and the following codings)}
\usage{
burninHMC(samples, n_burnin, n_thin = 1L)
}
\arguments{
\item{samples}{cube with final draws}

\item{n_burnin}{is the number of draws to discard after running the chains}

\item{n_thin}{is the number of draws to be skipped. For instance, if n_thin = 10, draws will be selected every 10 generated draws: 1, 11, 21, 31, ...}
}
\value{
cube with selected draws
}
\description{
This function performs the burn-in and the thinning at the end of the HMC
}
