% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_m}
\alias{compute_m}
\title{compute_m (one of the most important functions to compute quantity m, see paper)}
\usage{
compute_m(index, edgelist, U, S)
}
\arguments{
\item{index}{get_events_index output}

\item{edgelist}{(this argument is only used to define a dimension, which is the number of events, then it can be omitted)}

\item{U}{matrix of unique vectors}

\item{S}{array of statistics with dimensons [dyads*statistics*time]}
}
\value{
vector of m's
}
\description{
description of the function here
}
