% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_q}
\alias{compute_q}
\title{compute_q (one of the most important functions to compute quantity q, see paper)}
\usage{
compute_q(index, edgelist, U, S)
}
\arguments{
\item{index}{get_events_index output}

\item{edgelist}{(this argument is only used to define a dimension, which is the number of events, then it can be omitted)}

\item{U}{matrix of unique vectors}

\item{S}{array of statistics with dimensons [dyads*statistics*time]}
}
\value{
vector of q's
}
\description{
description of the function here
}
