% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeOccurrencies}
\alias{computeOccurrencies}
\title{computeOccurrencies}
\usage{
computeOccurrencies(edgelist, risksetMatrix, M, unique_vectors_stats, stats)
}
\arguments{
\item{edgelist}{is the preprocessed edgelist dataframe with information about [time,sender,receiver,type,weight] by row.}

\item{risksetMatrix}{matrix object inside the output list of the preprocessed relational event history.}

\item{M}{number of observed relational events.}

\item{unique_vectors_stats}{matrix of unique vectors of statistics (output of getUniqueVectors()).}

\item{stats}{array of statistics with dimensons [D*U*M]}
}
\value{
vector of q's
}
\description{
A function to compute how many times each of the unique vector of statistics returned by getUniqueVectors() occurred in the network (as in contributing to the hazard in the likelihood). This function is based on the result shown by the Appendix C in the paper 'Hierarchical models for relational event sequences', DuBois et al. 2013 (pp. 308-309).
}
