% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{computeTimes}
\alias{computeTimes}
\title{computeTimes}
\usage{
computeTimes(unique_vectors_stats, M, stats, intereventTime)
}
\arguments{
\item{unique_vectors_stats}{matrix of unique vectors of statistics (output of getUniqueVectors()).}

\item{M}{number of observed relational events.}

\item{stats}{array of statistics with dimensons [D*U*M].}

\item{intereventTime}{vector of time differences between two subsequent time points (i.d., waiting time between t[m] and t[m-1]).}
}
\value{
vector of sum of interevent times per each unique_vector_stats element
}
\description{
A function to compute the sum of interevent times for those vector of statistics that occurre more than once (output of getUniqueVectors()). This function is based on the result shown by the Appendix C in the paper 'Hierarchical models for relational event sequences', DuBois et al. 2013 (pp. 308-309).
}
