% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertEdgelist}
\alias{convertEdgelist}
\title{convertEdgelist}
\usage{
convertEdgelist(edgelist, actorsDictionary, typesDictionary, M)
}
\arguments{
\item{edgelist}{is the input data frame with information about [time,sender,receiver,type,weight] by row.}

\item{actorsDictionary}{dictionary of actors names (input string name = integer id)}

\item{typesDicitonary}{dictionary of event types (input string name = integer id)}
}
\value{
cube of possible combination [sender,receiver,type]: the cell value is the column index in the rehBinary matrix
}
\description{
convertEdgelist
}
