% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convertInputREH}
\alias{convertInputREH}
\title{convertInputREH}
\usage{
convertInputREH(edgelist, riskset, actorsDictionary, typesDictionary, M)
}
\arguments{
\item{edgelist}{is the input data frame with information about [time,sender,receiver,type,weight] by row.}

\item{riskset}{riskset list with old actors sitring names.}

\item{actorsDictionary}{dictionary of actors names (input string name = integer id)}

\item{M}{number of observed relational events}

\item{typesDicitonary}{dictionary of event types (input string name = integer id)}
}
\value{
cube of possible combination [sender,receiver,type]: the cell value is the column index in the rehBinary matrix
}
\description{
convertInputREH
}
