% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convtoASCII}
\alias{convtoASCII}
\title{convtoASCII}
\usage{
convtoASCII(rehBinary)
}
\arguments{
\item{rehBinary}{matrix of integers 0/1 (dimensions are rows = M [number of events], columns = D [number of dyads])}
}
\value{
vector of char per each dyad (list of length D)
}
\description{
converting rehBinary to characters using 8bit and
}
