% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cube_to_matrix}
\alias{cube_to_matrix}
\title{cube_to_matrix (a function to transform a cube to a matrix)}
\usage{
cube_to_matrix(S)
}
\arguments{
\item{S}{cube structure to rearrange in a matrix of dimension [N(N-1)M*statistics]}
}
\value{
matrix
}
\description{
description of the function here
}
