% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{experimental}
\alias{experimental}
\title{experimental}
\usage{
experimental(pars, stats)
}
\arguments{
\item{pars}{parameters}

\item{stats}{array of statistics}
}
\value{
string
}
\description{
blah blah blah
}
