% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{experimental_function}
\alias{experimental_function}
\title{experimental_function (where to try out specific operations at C++ level)}
\usage{
experimental_function(x)
}
\arguments{
\item{x}{integer value}
}
\value{
matrix
}
\description{
the experimental function has no description
}
