% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GDADAMAX}
\alias{GDADAMAX}
\title{GDADAMAX}
\usage{
GDADAMAX(
  pars,
  stats,
  actor1,
  actor2,
  dyad,
  omit_dyad,
  interevent_time,
  model,
  ordinal = FALSE,
  senderRate = TRUE,
  gradient = TRUE,
  hessian = FALSE,
  N = NULL,
  C = NULL,
  D = NULL,
  ncores = 1L,
  epochs = 1000L,
  learning_rate = 0.002,
  beta1 = 0.9,
  beta2 = 0.999,
  epsilon = 0.01
)
}
\arguments{
\item{pars}{parameters}

\item{stats}{array of statistics}

\item{actor1}{vector of actor1's (column reh$edgelist$actor1)}

\item{actor2}{vector of actor2's (column reh$edgelist$actor2)}

\item{dyad}{vector of dyad (from the attribute attr(remify,"dyad"))}

\item{omit_dyad}{is a list of two objects: vector "time" and matrix "riskset". Two object for handling changing risksets. NULL if no change is defined}

\item{interevent_time}{vector of interevent times (inside the reh object)}

\item{model}{either "actor" or "tie" model}

\item{ordinal}{whether to use(TRUE) the ordinal likelihood or not (FALSE) then using the interval likelihood}

\item{senderRate}{boolean true/false (it is used only when model = "actor") indicates if to estimate the senderRate model (true) or the ReceiverChoice model (false)}

\item{gradient}{boolean true/false whether to return gradient value}

\item{hessian}{boolean true/false whether to return hessian value}

\item{N}{number of actors. This argument is used only in the ReceiverChoice likelihood (model = "actor")}

\item{C}{number of event types}

\item{D}{number of dyads}

\item{ncores}{number of threads to use for the parallelization}

\item{epochs}{number of epochs}

\item{learning_rate}{learning rate}

\item{beta1}{hyperparameter beta1}

\item{beta2}{hyperparameter beta2}

\item{epsilon}{hyperparameter eta}
}
\value{
optimization with GDADAM
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values
}
