% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_events_index}
\alias{get_events_index}
\title{get_events_index (to remove, since a similar function will be written in reh.h)}
\usage{
get_events_index(edgelist, riskset)
}
\arguments{
\item{edgelist}{matrix of time sender and receiver}

\item{riskset}{matrix of possible dyads}
}
\value{
vector of indices (indicating the dyad occurred at each time point)
}
\description{
description of the function here
}
