% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_unique_vectors}
\alias{get_unique_vectors}
\title{get_unique_vectors (a function to )}
\usage{
get_unique_vectors(A)
}
\arguments{
\item{A}{matrix of statistics per dyad and per time point}
}
\value{
matrix with only unique vectors
}
\description{
description of the function here
}
