% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBinaryREH}
\alias{getBinaryREH}
\title{getBinaryREH (a function that returns a utility matrix used in optimization algorithms)}
\usage{
getBinaryREH(edgelist, riskset, risksetCube, M, D)
}
\arguments{
\item{edgelist}{edgelist converted according to actorID and typeID}

\item{riskset}{riskset list converted according to actorID and typeID}

\item{risksetCube}{arma::cube object [N*N*C] where the cell value returns the column index to use in the outBinaryREH}

\item{M}{number of observed relational events}

\item{D}{number of possible dyads (accounting for event types as well)}
}
\value{
utility matrix per row 0 if the event could happen but didn't, 1 if the event happend, -1 if the event couldn't occur
}
\description{
getBinaryREH (a function that returns a utility matrix used in optimization algorithms)
}
