% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRiskset}
\alias{getRiskset}
\title{getRiskset (obtain permutations of actors' ids and event types).}
\usage{
getRiskset(actorsID, typesID, N, C)
}
\arguments{
\item{actorsID}{vector of actors' id's.}

\item{typesID}{vector of types' id's.}

\item{N}{number of actors in the dataset.}

\item{C}{number of event types}
}
\value{
matrix of possible dyadic events.
}
\description{
getRiskset (obtain permutations of actors' ids and event types).
}
