% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRisksetCube}
\alias{getRisksetCube}
\title{getRisksetCube}
\usage{
getRisksetCube(risksetMatrix, N, C)
}
\arguments{
\item{risksetMatrix}{output of getRiskset() function}

\item{N}{number of actors in the dataset.}

\item{C}{number of event types}
}
\value{
cube of possible combination [sender,receiver,type]: the cell value is the column index in the rehBinary matrix
}
\description{
getRisksetCube
}
