% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getRisksetMatrix}
\alias{getRisksetMatrix}
\title{getRisksetMatrix (obtain permutations of actors' ids and event types).}
\usage{
getRisksetMatrix(actorID, typeID, N, C)
}
\arguments{
\item{actorID}{vector of actors' id's.}

\item{typeID}{vector of types' id's.}

\item{N}{number of actors in the dataset.}

\item{C}{number of event types}
}
\value{
matrix of possible dyadic events.
}
\description{
getRisksetMatrix (obtain permutations of actors' ids and event types).
}
