% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getUniqueVectors}
\alias{getUniqueVectors}
\title{getUniqueVectors}
\usage{
getUniqueVectors(stats)
}
\arguments{
\item{stats}{cube structure of dimensions [D*U*M] filled with statistics values.}
}
\value{
matrix with only unique vectors of statistics with dimensions [R*U]
}
\description{
A function to retrieve only the unique vectors of statistics observed throught times points and dyads. This function is based on the result shown by the Appendix C in the paper 'Hierarchical models for relational event sequences', DuBois et al. 2013 (pp. 308-309).
}
