% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HMC}
\alias{HMC}
\title{HMC}
\usage{
HMC(
  pars_init,
  nsim,
  nchains,
  burnin,
  meanPrior,
  sigmaPrior,
  stats,
  actor1,
  actor2,
  dyad,
  omit_dyad,
  interevent_time,
  model,
  ordinal = FALSE,
  ncores = 1L,
  senderRate = TRUE,
  N = NULL,
  C = NULL,
  D = NULL,
  thin = 1L,
  L = 100L,
  epsilon = 0.01
)
}
\arguments{
\item{pars_init}{is a matrix of dimensions U x nchains where for each column (chain) a random vector of initial values for the parameter is supplied.}

\item{nsim}{is the number of samples from the posterior that have to be generated.}

\item{nchains}{number of chains of length nsim}

\item{burnin}{is the number of draws to discard after running the chains}

\item{meanPrior}{is a vector of prior means with the same dimension as the vector of parameters}

\item{sigmaPrior}{is a matrix, I have been using a diagonal matrix here with the same dimension as the vector os parameters}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{actor1}{vector of actor1's (column reh$edgelist$actor1)}

\item{actor2}{vector of actor2's (column reh$edgelist$actor2)}

\item{dyad}{vector of dyad (from the attribute attr(remify,"dyad"))}

\item{omit_dyad}{is a list of two objects: vector "time" and matrix "riskset". Two object for handling changing risksets. NULL if no change is defined//' @param interevent_time the time difference between the current time point and the previous event time.//' @param}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{model}{either "actor" or "tie" model}

\item{ordinal}{logic TRUE/FALSE}

\item{ncores}{number of threads for parallel computing (default = 1)}

\item{senderRate}{boolean true/false (it is used only when model = "actor") indicates if to estimate the senderRate model (true) or the ReceiverChoice model (false)}

\item{N}{number of actors. This argument is used only in the ReceiverChoice likelihood (model = "actor")}

\item{C}{number of event types}

\item{D}{number of dyads}

\item{thin}{is the number of draws to be skipped. For instance, if thin = 10, draws will be selected every 10 generated draws: 1, 11, 21, 31, ...}

\item{L}{number of leapfrogs. Default (and recommended) value is 100.}

\item{epsilon}{size of the leapfrog. Default value is 1e-02.}
}
\value{
posterior draws
}
\description{
This function performs the Hamiltonian Monte Carlo
}
