% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logLike}
\alias{logLike}
\title{logLike (computing loglikelihood value)}
\usage{
logLike(beta, U, q, m)
}
\arguments{
\item{beta}{vector of parameters' value where to compute the value of the loglikelihood}

\item{U}{matrix of unique vectors}

\item{q}{vector computed with compute_q}

\item{m}{vector computed with compute_m}
}
\value{
vector of q's
}
\description{
description of the function here
}
