% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logPostHMC}
\alias{logPostHMC}
\title{logPostHMC}
\usage{
logPostHMC(
  meanPrior,
  sigmaPrior,
  pars,
  stats,
  event_binary,
  interevent_time,
  times_r,
  occurrencies_r,
  unique_vectors_stats,
  fast
)
}
\arguments{
\item{meanPrior}{is a vector of prior means with the same dimension as the vector of parameters}

\item{sigmaPrior}{is a matrix, I have been using a diagonal matrix here with the same dimension as the vector os parameters}

\item{pars}{is a vector of parameters (note: the order must be aligned with the column order in 'stats')}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{event_binary}{is a matrix [M*D] of 1/0/-1 : 1 indicating the observed dyad and 0 (-1) the non observed dyads that could have (have not) occurred.}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{times_r}{used in the fast approach}

\item{occurrencies_r}{used in the fast approach}

\item{unique_vectors_stats}{used in the fast approach}

\item{fast}{boolean true/false whether to run the fast approach or not}
}
\value{
value of log-posterior density
}
\description{
This function calculates the value of the log-posterior density given the loglikelihood and the log-prior density
}
