% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nllik}
\alias{nllik}
\title{nllik (Negative Log-Likelihood of REM)}
\usage{
nllik(pars, stats, event_binary, interevent_time, threads)
}
\arguments{
\item{pars}{is a vector of parameters (note: the order must be the same as the column order in 'stats') at which to calculate the likelihood value}

\item{stats}{is a cube of dimensions n_dyads*variables*M with statistics of interest by column and dyads by row.}

\item{event_binary}{is a matrix of ones and zeros of dimensions M*n_dyads : 1 indicating the observed dyad and 0 the non-observed dyads.}

\item{interevent_time}{the vector of time differences between the current time point and the previous event time (note: interevent time at t_1 is t_1-0=t_1, since t_0 = 0).}

\item{threads}{}
}
\value{
negative log likelihood value
}
\description{
nllik (Negative Log-Likelihood of REM)
}
