% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{print.remstimate}
\alias{print.remstimate}
\title{print.remstimate}
\usage{
\method{print}{remstimate}(remstimate, ...)
}
\arguments{
\item{remstimate}{is a \code{remstimate} object}

\item{...}{further arguments to be passed.}
}
\description{
A function that prints out the estimates of the selected optimization method.
}
