% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{print.summary.remstimate}
\alias{print.summary.remstimate}
\title{print.summary.remstimate}
\usage{
\method{print}{summary.remstimate}(x, ...)
}
\arguments{
\item{x}{is a \code{summary.remstimate} object}

\item{...}{further arguments to be passed to the 'print.summary' method}
}
\description{
A function that prints out a summary of a 'remstimate' object.
}
