% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/reh.R
\name{reh}
\alias{reh}
\title{reh (a function to pre-process the REH input)}
\usage{
reh(edgelist, riskset, covariates)

reh(edgelist, riskset, covariates)
}
\arguments{
\item{edgelist}{is a dataframe of relational events sorted by time: [time,sender,receiver,type,weight]}

\item{riskset}{is a list of length the number of events, each object a matrix with unobserved dyads (using actors string names)}

\item{covariates}{list of covariates to be provided according to the input structure working with 'remstats'}

\item{actorsDictionary}{}
}
\value{
list of objects useful for the analysis of the REH (read more here)

object list (the function saves also the output of optim)
}
\description{
description of remstimate::reh here (most likeliy the function won't be exported as a function in R but only intended to be used internally, since the
conversion to integer id's and all that follows can be a kind of pre-processing difficult to digest and not so intuitive)

A function that returns maximum likelihood estimates of REM (interval timing only) by using the optim function
}
