% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivatives}
\alias{remDerivatives}
\title{remDerivatives}
\usage{
remDerivatives(
  pars,
  stats,
  actor1,
  actor2,
  dyad,
  omit_dyad,
  interevent_time,
  model,
  ordinal = FALSE,
  ncores = 1L,
  gradient = TRUE,
  hessian = TRUE,
  senderRate = TRUE,
  N = NULL,
  C = NULL,
  D = NULL
)
}
\arguments{
\item{pars}{is a vector of parameters (note: the order must be aligned with the column order in 'stats')}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{actor1}{vector of actor1's (column reh$edgelist$actor1)}

\item{actor2}{vector of actor2's (column reh$edgelist$actor2)}

\item{dyad}{vector of dyad (from the attribute attr(remify,"dyad"))}

\item{omit_dyad}{is a list of two objects: vector "time" and matrix "riskset". Two object for handling changing risksets. NULL if no change is defined}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{model}{either "actor" or "tie" model}

\item{ordinal}{whether to use(TRUE) the ordinal likelihood or not (FALSE) then using the interval likelihood}

\item{ncores}{number of threads to use for the parallelization}

\item{gradient}{boolean true/false whether to return gradient value}

\item{hessian}{boolean true/false whether to return hessian value}

\item{senderRate}{boolean true/false (it is used only when model = "actor") indicates if to estimate the senderRate model (true) or the ReceiverChoice model (false)}

\item{N}{number of actors. This argument is used only in the ReceiverChoice likelihood (model = "actor")}

\item{C}{number of event types}

\item{D}{number of dyads}
}
\value{
list of values: loglik, gradient, hessian
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values
}
