% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivativesFast}
\alias{remDerivativesFast}
\title{remDerivativesFast (a function that returns a list of 0th/1st/2nd order derivatives of loglikelihood evaluated in pars)}
\usage{
remDerivativesFast(pars, times_r, occurrencies_r, unique_vectors_stats)
}
\arguments{
\item{pars}{vector of parameters}

\item{times_r}{former m}

\item{occurrencies_r}{former q}

\item{unique_vectors_stats}{former U}
}
\value{
list of value/gradient/hessian in pars
}
\description{
description of the function here
}
