% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivativesReceiverChoice}
\alias{remDerivativesReceiverChoice}
\title{remDerivativesReceiverChoice}
\usage{
remDerivativesReceiverChoice(
  pars,
  stats,
  actor1,
  actor2,
  omit_dyad,
  interevent_time,
  N,
  C,
  D,
  gradient = TRUE,
  hessian = TRUE
)
}
\arguments{
\item{pars}{is a vector of parameters}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{actor1}{vector of actor1's (column reh$edgelist$actor1)}

\item{actor2}{vector of actor2's (column reh$edgelist$actor2)}

\item{omit_dyad, }{list object that takes care of the dynamic rikset (if defined)}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{N}{the number of actors}

\item{C}{number of event types}

\item{D}{number of dyads}

\item{gradient}{boolean true/false whether to return gradient value}

\item{hessian}{boolean true/false whether to return hessian value}
}
\value{
list of values: loglik, grad, fisher
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values for estimating the receiver choice parameters for the actor oriented model
}
