% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivativesSenderRates}
\alias{remDerivativesSenderRates}
\title{remDerivativesSenderRates}
\usage{
remDerivativesSenderRates(
  pars,
  stats,
  edgelist,
  omit_dyad,
  interevent_time,
  C,
  D,
  ordinal = FALSE
)
}
\arguments{
\item{pars}{is a vector of parameters}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions N*U with statistics of interest by column and senders by row.}

\item{edgelist}{is a matrix [M*3] of [time/dyad/weight]}

\item{omit_dyad}{is a list of two objects: vector "time" and matrix "riskset". Two object for handling changing risksets. NULL if no change is defined}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{C}{number of event types}

\item{D}{number of dyads}

\item{ordinal}{boolean, true if the likelihood to use is the ordinal one, interval otherwise}
}
\value{
list of values: loglik, grad, fisher information
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values for estimating the sender rate parameters for the actor oriented model
}
