% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate-package.R
\docType{package}
\name{remstimate-package}
\alias{remstimate-package}
\title{remstimate: Optimization Frameworks for Tie-Oriented and Actor-Oriented Relational Event Models}
\description{
Tools for fitting, diagnozing, and analyzing relational event models. There is support for tie-oriented modeling (Butts, C., 2008, \doi{10.1111/j.1467-9531.2008.00203.x}) and an actor-oriented modeling framework (Stadtfeld, C., & Block, P., 2017, \doi{10.15195/v4.a14}). The package accommodates both frequentist and Bayesian approaches. There is support for modeling of events with a duration, for random effects models (e.g., frailty relational event models), discrete mixture models (e.g., dyadic latent class modeling), and penalized approaches (e.g., using the lasso).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tilburgnetworkgroup.github.io/remstimate/}
  \item Report bugs at \url{https://github.com/TilburgNetworkGroup/remstimate/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Arena \email{g.arena@uva.nl} (\href{https://orcid.org/0000-0001-5204-3326}{ORCID})

Authors:
\itemize{
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu}
  \item Rumana Lakdawala
  \item Fabio Generoso Vieira
}

Other contributors:
\itemize{
  \item Marlyne Meijerink-Bosman [contributor]
  \item Diana Karimova [contributor]
  \item Mahdi Shafiee Kamalabad [contributor]
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
