% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{summary.remstimate}
\alias{summary.remstimate}
\title{summary.remstimate}
\usage{
\method{summary}{remstimate}(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the 'summary' method}
}
\description{
A function that arranges a summary of a 'remstimate' object
}
