% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_ic.R
\name{WAIC}
\alias{WAIC}
\alias{WAIC.remstimate}
\title{WAIC}
\usage{
WAIC(object, ...)

\method{WAIC}{remstimate}(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object.}

\item{...}{further arguments to be passed to the 'WAIC' method.}
}
\value{
WAIC value of a 'remstimate' object.
}
\description{
A function that returns the WAIC (Watanabe-Akaike's Information Criterion) value in a 'remstimate' object.
}
\section{Methods (by class)}{
\itemize{
\item \code{WAIC(remstimate)}: WAIC (Watanabe-Akaike's Information Criterion) value of a 'remstimate' object

}}
\examples{

# No examples available at the moment

}
