% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{diagnostics}
\alias{diagnostics}
\alias{diagnostics.remstimate}
\title{Compute the diagnostics of a \code{remstimate} object}
\usage{
diagnostics(object, reh, stats, ...)

\method{diagnostics}{remstimate}(object, reh, stats, top_pct = 0.05, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object.}

\item{reh}{is a \code{remify} object, the same used for the 'remstimate' object.}

\item{stats}{is a \code{remstats} object, the same used for the 'remstimate' object.}

\item{...}{further arguments to be passed to the 'diagnostics' method.}

\item{top_pct}{numeric scalar in (0,1): threshold for the top-percentage recall summary (default 0.05).}
}
\description{
Compute the diagnostics of a \code{remstimate} object
}
\section{Methods (by class)}{
\itemize{
\item \code{diagnostics(remstimate)}: diagnostics of a 'remstimate' object

}}
