% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GD}
\alias{GD}
\title{GD}
\usage{
GD(
  pars,
  stats,
  edgelist,
  omit_dyad,
  interevent_time,
  model,
  ordinal = FALSE,
  ncores = 1L,
  epochs = 200L,
  learning_rate = 0.001
)
}
\arguments{
\item{pars}{parameters}

\item{stats}{array of statistics}

\item{edgelist}{is a matrix [M*3] of [time/dyad/weight]}

\item{omit_dyad}{is a list of two objects: vector "time" and matrix "riskset". Two object for handling changing risksets. NULL if no change is defined}

\item{interevent_time}{vector of interevent times (inside the reh object)}

\item{model}{either "actor" or "tie" model}

\item{ordinal}{whether to use(TRUE) the ordinal likelihood or not (FALSE) then using the interval likelihood}

\item{ncores}{number of threads to use for the parallelization}

\item{epochs}{number of epochs}

\item{learning_rate}{learning rate}
}
\value{
optimization with Gradient Descent algorithm
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values
}
