% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GD}
\alias{GD}
\title{GD}
\usage{
GD(
  pars,
  stats,
  event_binary,
  interevent_time,
  times_r,
  occurrencies_r,
  unique_vectors_stats,
  fast = FALSE,
  epochs = 200L,
  learning_rate = 1e-04
)
}
\arguments{
\item{pars}{parameters}

\item{stats}{array of statistics}

\item{event_binary}{rehBinary (inside the reh object)}

\item{interevent_time}{vector of interevent times (inside the reh object)}

\item{times_r}{used in the fast approach}

\item{occurrencies_r}{used in the fast approach}

\item{unique_vectors_stats}{used in the fast approach}

\item{fast}{TRUE/FALSE whether to perform the fast approach or not}

\item{epochs}{number of epochs}

\item{learning_rate}{learning rate}
}
\value{

}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values
}
