% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HMC}
\alias{HMC}
\title{HMC}
\usage{
HMC(
  pars_init,
  n_iters,
  n_chains,
  n_burnin,
  meanPrior,
  sigmaPrior,
  stats,
  event_binary,
  interevent_time,
  times_r,
  occurrencies_r,
  unique_vectors_stats,
  fast = FALSE,
  n_thin = 1L,
  L = 100L,
  epsilon = 0.01,
  n_threads = 1L
)
}
\arguments{
\item{pars_init}{is a matrix of dimensions U x n_chains where for each column (chain) a random vector of initial values for the parameter is supplied.}

\item{n_iters}{is the number of samples from the posterior that have to be generated.}

\item{n_chains}{number of chains of length n_iters}

\item{n_burnin}{is the number of draws to discard after running the chains}

\item{meanPrior}{is a vector of prior means with the same dimension as the vector of parameters}

\item{sigmaPrior}{is a matrix, I have been using a diagonal matrix here with the same dimension as the vector os parameters}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{event_binary}{is a matrix [M*D] of 1/0/-1 : 1 indicating the observed dyad and 0 (-1) the non observed dyads that could have (have not) occurred.}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{times_r}{used in the fast approach}

\item{occurrencies_r}{used in the fast approach}

\item{unique_vectors_stats}{used in the fast approach}

\item{fast}{boolean TRUE/FALSE whether to run the fast approach or not (default = FALSE)}

\item{n_thin}{is the number of draws to be skipped. For instance, if n_thin = 10, draws will be selected every 10 generated draws: 1, 11, 21, 31, ...}

\item{L}{number of leapfrogs. Default (and recommended) value is 100.}

\item{epsilon}{size of the leapfrog. Default value is 1e-02.}

\item{n_threads}{number of threads for parallel computing (default = 1)}

\item{pars}{is a vector of parameters (note: the order must be aligned with the column order in 'stats')}
}
\value{
matrix with posterior draws
}
\description{
This function performs the Hamiltonian Monte Carlo
}
