% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HMC}
\alias{HMC}
\title{HMC}
\usage{
HMC(
  pars_init,
  nsim,
  nchains,
  burnin,
  meanPrior,
  sigmaPrior,
  stats,
  edgelist,
  omit_dyad,
  interevent_time,
  model,
  ordinal = FALSE,
  ncores = 1L,
  fast = FALSE,
  thin = 1L,
  L = 100L,
  epsilon = 0.01
)
}
\arguments{
\item{pars_init}{is a matrix of dimensions U x nchains where for each column (chain) a random vector of initial values for the parameter is supplied.}

\item{nsim}{is the number of samples from the posterior that have to be generated.}

\item{nchains}{number of chains of length nsim}

\item{burnin}{is the number of draws to discard after running the chains}

\item{meanPrior}{is a vector of prior means with the same dimension as the vector of parameters}

\item{sigmaPrior}{is a matrix, I have been using a diagonal matrix here with the same dimension as the vector os parameters}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{edgelist}{is a matrix [M*3] of [time/dyad/weight]}

\item{omit_dyad}{is a list of two objects: vector "time" and matrix "riskset". Two object for handling changing risksets. NULL if no change is defined//' @param interevent_time the time difference between the current time point and the previous event time.//' @param}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{model}{either "actor" or "tie" model}

\item{ordinal}{logic TRUE/FALSE}

\item{ncores}{number of threads for parallel computing (default = 1)}

\item{fast}{boolean TRUE/FALSE whether to run the fast approach or not (default = FALSE)}

\item{thin}{is the number of draws to be skipped. For instance, if thin = 10, draws will be selected every 10 generated draws: 1, 11, 21, 31, ...}

\item{L}{number of leapfrogs. Default (and recommended) value is 100.}

\item{epsilon}{size of the leapfrog. Default value is 1e-02.}
}
\value{
posterior draws
}
\description{
This function performs the Hamiltonian Monte Carlo
}
