% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logPostHMC}
\alias{logPostHMC}
\title{logPostHMC}
\usage{
logPostHMC(
  meanPrior,
  sigmaPrior,
  pars,
  stats,
  edgelist,
  omit_dyad,
  interevent_time,
  model,
  ordinal = FALSE,
  ncores = 1L,
  fast = FALSE
)
}
\arguments{
\item{meanPrior}{is a vector of prior means with the same dimension as the vector of parameters}

\item{sigmaPrior}{is a matrix, I have been using a diagonal matrix here with the same dimension as the vector os parameters}

\item{pars}{is a vector of parameters (note: the order must be aligned with the column order in 'stats')}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{edgelist}{is a matrix [M*3] of [time/dyad/weight]}

\item{omit_dyad}{is a list of two objects: vector "time" and matrix "riskset". Two object for handling changing risksets. NULL if no change is defined//' @param interevent_time the time difference between the current time point and the previous event time.}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{model}{either "actor" or "tie" model}

\item{ordinal}{whether to use(TRUE) the ordinal likelihood or not (FALSE) then using the interval likelihood}

\item{ncores}{number of threads to use for the parallelization}

\item{fast}{boolean true/false whether to run the fast approach or not}
}
\value{
value of log-posterior density
}
\description{
This function calculates the value of the log-posterior density given the loglikelihood and the log-prior density
}
