% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{plot.remstimate}
\alias{plot.remstimate}
\title{plot.remstimate}
\usage{
\method{plot}{remstimate}(x, ...)
}
\arguments{
\item{x}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the 'predict' method depending on the estimator used}
}
\description{
A function that returns a plot of diagnostics given a 'remstimate' object and depending on the 'approach' attribute.
}
