% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{predict.remstimate}
\alias{predict.remstimate}
\title{predict.remstimate}
\usage{
\method{predict}{remstimate}(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the 'predict' method depending on the estimator used}
}
\description{
A function that returns predictions given a 'remstimate' object.
}
