% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{print.remstimate}
\alias{print.remstimate}
\title{print.remstimate}
\usage{
\method{print}{remstimate}(x, ...)
}
\arguments{
\item{x}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the print method}
}
\description{
A function that prints out the estimates returned by a 'remstimate' object.
}
