% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{print.summary.remstimate}
\alias{print.summary.remstimate}
\title{print.summary.remstimate()}
\usage{
\method{print}{summary.remstimate}(x, ...)
}
\arguments{
\item{x}{is a \code{summary.remstimate} object}

\item{...}{further arguments to be passed.}
}
\description{
A function that prints out a summary of the optimization result (codings are based on the structure of summary.lm() and summary.glm() from package 'stats')
}
