% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{reh}
\alias{reh}
\title{reh}
\usage{
reh(
  edgelist,
  riskset = replicate(NROW(edgelist), NaN, simplify = FALSE),
  covariates = list(a = NULL)
)
}
\arguments{
\item{edgelist}{a a dataframe of relational events sorted by time: [time,sender,receiver,type,weight]}

\item{riskset}{is a list of length equal to the number of events, each object a matrix with unobserved dyads (using actors string names)}

\item{covariates}{list of covariates to be provided according to the input structure working with 'remstats'}
}
\value{
object list (the function saves also the output of optim)
}
\description{
A function that returns a 'reh' S3 object
}
