% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{reh}
\alias{reh}
\title{reh (a function for preprocessing data)}
\usage{
reh(edgelist, riskset, covariates)
}
\arguments{
\item{edgelist}{is a dataframe of relational events sorted by time: [time,sender,receiver,type,weight]}

\item{riskset}{is a list of length the number of events, each object a matrix with unobserved dyads (using actors string names)}

\item{covariates}{list of covariates to be provided according to the input structure working with 'remstats'}
}
\value{
list of objects
}
\description{
reh (a function for preprocessing data)
}
