% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivatives}
\alias{remDerivatives}
\title{remDerivatives ( [OLD/NOT USED] function originally useful to switch between normal and fast routine for computing the likelihood)}
\usage{
remDerivatives(
  pars,
  stats,
  event_binary,
  interevent_time,
  model,
  ordinal = FALSE,
  ncores = 1L,
  fast = FALSE,
  gradient = TRUE,
  hessian = TRUE
)
}
\arguments{
\item{pars}{is a vector of parameters (note: the order must be aligned with the column order in 'stats')}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{event_binary}{is a matrix [M*D] of 1/0/-1 : 1 indicating the observed dyad and 0 (-1) the non observed dyads that could have (have not) occurred.}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{model}{either "actor" or "tie" model}

\item{ordinal}{whether to use(TRUE) the ordinal likelihood or not (FALSE) then using the interval likelihood}

\item{ncores}{number of threads to use for the parallelization}

\item{fast}{boolean true/false whether to run the fast approach or not}

\item{gradient}{boolean true/false whether to return gradient value}

\item{hessian}{boolean true/false whether to return hessian value}
}
\value{
list of values: loglik, gradient, hessian
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values
}
