% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivativesFast}
\alias{remDerivativesFast}
\title{remDerivativesFast}
\usage{
remDerivativesFast(
  pars,
  times_r,
  occurrencies_r,
  unique_vectors_stats,
  gradient,
  hessian
)
}
\arguments{
\item{pars}{vector of parameters}

\item{times_r}{former m}

\item{occurrencies_r}{former q}

\item{unique_vectors_stats}{former U}

\item{gradient}{boolean}

\item{hessian}{boolean}
}
\value{
list of value/gradient/hessian in pars
}
\description{
description of the function here
}
