% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivativesReceiverChoice}
\alias{remDerivativesReceiverChoice}
\title{remDerivativesReceiverChoice}
\usage{
remDerivativesReceiverChoice(
  pars,
  stats,
  edgelist,
  omit_dyad,
  interevent_time,
  directed,
  N,
  C,
  D
)
}
\arguments{
\item{pars}{is a vector of parameters}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{edgelist, }{output from remify, (note: indices of the actors must start from 0)}

\item{omit_dyad, }{list object that takes care of the dynamic rikset (if defined)}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{directed, }{boolean TRUE/FALSE if the network is directed or not}

\item{N, }{the number of actors}

\item{C, }{number of event types}
}
\value{
list of values: loglik, grad, fisher
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values for estimating the receiver choice parameters for the actor oriented model
}
