% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivativesSenderRates}
\alias{remDerivativesSenderRates}
\title{remDerivativesSenderRates}
\usage{
remDerivativesSenderRates(
  pars,
  stats,
  risksetCube,
  event_binary,
  interevent_time,
  edgelist
)
}
\arguments{
\item{pars}{is a vector of parameters}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions N*U with statistics of interest by column and senders by row.}

\item{event_binary}{is a matrix [M*D] of 1/0/-1 : 1 indicating the observed dyad and 0 (-1) the non observed dyads that could have (have not) occurred.}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{edgelist, }{output from remify, (note: indices of the actors must start from 0)}
}
\value{
list of values: loglik, grad, fisher
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values for estimating the sender rate parameters for the actor oriented model
}
