% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{remDerivativesStandard}
\alias{remDerivativesStandard}
\title{remDerivativesStandard}
\usage{
remDerivativesStandard(
  pars,
  stats,
  event_binary,
  interevent_time,
  ordinal = FALSE,
  ncores = 1L,
  gradient = TRUE,
  hessian = TRUE
)
}
\arguments{
\item{pars}{is a vector of parameters (note: the order must be aligned with the column order in 'stats')}

\item{stats}{is cube of M slices. Each slice is a matrix of dimensions D*U with statistics of interest by column and dyads by row.}

\item{event_binary}{is a matrix [M*D] of 1/0/-1 : 1 indicating the observed dyad and 0 (-1) the non observed dyads that could have (have not) occurred.}

\item{interevent_time}{the time difference between the current time point and the previous event time.}

\item{ordinal}{boolean that indicate whether to use the ordinal or interval timing likelihood}

\item{ncores}{integer referring to the number of threads for the parallelization}

\item{gradient}{boolean true/false whether to return gradient value}

\item{hessian}{boolean true/false whether to return hessian value}
}
\value{
list of values: loglik, gradient, hessian
}
\description{
function that returns a list as an output with loglikelihood/gradient/hessian values at specific parameters' values
}
