% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate-package.R
\docType{package}
\name{remstimate-package}
\alias{remstimate-package}
\title{remstimate: Optimization Frameworks for Tie-Oriented and Actor-Oriented Relational Event Models}
\description{
Tools for fitting, diagnosing, and analyzing tie-oriented and actor-oriented relational event models, under both frequentist and Bayesian approaches. The package supports tie-oriented modeling (Butts, 2008, \doi{10.1111/j.1467-9531.2008.00203.x}) and an actor-oriented modeling framework (Stadtfeld et al., 2017, \doi{10.15195/v4.a14}), with additional model diagnostics and goodness-of-fit tools. Interfaces to estimation backends provide a range of extensions: random-effects (frailty) relational event models capturing sender, receiver, and dyadic heterogeneity (Juozaitiene & Wit 2024, \doi{10.1007/s11336-024-09952-x}; Mulder & Hoff, 2024, \doi{10.1214/24-AOAS1885}), finite mixture and dyadic latent class models for unobserved dyadic heterogeneity (Lakdawala et al., 2026, \doi{10.1016/j.socnet.2026.06.006}), penalized estimation via the lasso, ridge, and elastic net (Tibshirani, R., 1996, \doi{10.1111/j.2517-6161.1996.tb02080.x}; Karimova et al., 2023, \doi{10.1016/j.socnet.2023.02.006}), and approximate Bayesian regularization (Karimova et al., 2025, \doi{10.1016/j.jmp.2025.102925}). Modeling of events with a duration is also supported (Lakdawala et al., 2026, \doi{10.48550/arXiv.2602.21000}) and moving window relational event models (Mulder & Leenders, 2019, \doi{10.1016/j.chaos.2018.11.027}; Meijerink et al., 2023, \doi{10.1371/journal.pone.0272309}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tilburgnetworkgroup.github.io/remstimate/}
  \item Report bugs at \url{https://github.com/TilburgNetworkGroup/remstimate/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Arena \email{g.arena@uva.nl} (\href{https://orcid.org/0000-0001-5204-3326}{ORCID})

Authors:
\itemize{
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu}
  \item Rumana Lakdawala
  \item Fabio Generoso Vieira
}

Other contributors:
\itemize{
  \item Marlyne Meijerink-Bosman [contributor]
  \item Diana Karimova [contributor]
  \item Mahdi Shafiee Kamalabad [contributor]
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
