% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate-package.R
\docType{package}
\name{remstimate-package}
\alias{remstimate}
\alias{remstimate-package}
\title{remstimate: optimization tools for tie-oriented and actor-oriented relational event models}
\description{
This package offers a set of tools for the optimization of the likelihood in a relational event modeling framework (initial objective: to compare time performance with relevent::rem function and to improve them).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TilburgNetworkGroup/remstimate}
  \item Report bugs at \url{https://github.com/TilburgNetworkGroup/remstimate/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Arena \email{g.arena@tilburguniversity.edu}

Authors:
\itemize{
  \item Rumana Lakdawala \email{r.j.lakdawala@tilburguniversity.edu}
  \item Marlyne Meijerink-Bosman \email{m.l.meijerink@tilburguniversity.edu}
  \item Diana Karimova \email{d.karimova@tilburguniversity.edu}
  \item Mahdi Shafiee Kamalabad \email{m.shafiee@tilburguniversity.edu}
  \item Fabio Generoso Vieira \email{f.v.generosovieira@tilburguniversity.edu}
}

Other contributors:
\itemize{
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu} [contributor]
}

}
\keyword{.rd}
\keyword{be}
\keyword{create}
\keyword{defined}
\keyword{documentation}
\keyword{file}
\keyword{functions}
\keyword{if}
\keyword{in}
\keyword{internal}
\keyword{keywords}
\keyword{the}
\keyword{there}
\keyword{this}
\keyword{will}
\keyword{won't}
