% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate-package.R
\docType{package}
\name{remstimate-package}
\alias{remstimate-package}
\alias{_PACKAGE}
\title{remstimate: optimization tools for tie-oriented and actor-oriented relational event models}
\description{
This package offers a set of tools for the optimization of the likelihood in a relational event modeling framework. Frequentist as well as Bayesian approaches are available inside the package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TilburgNetworkGroup/remstimate}
  \item Report bugs at \url{https://github.com/TilburgNetworkGroup/remstimate/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Arena \email{g.arena@tilburguniversity.edu}

Authors:
\itemize{
  \item Rumana Lakdawala \email{r.j.lakdawala@tilburguniversity.edu}
  \item Fabio Generoso Vieira \email{f.v.generosovieira@tilburguniversity.edu}
}

Other contributors:
\itemize{
  \item Marlyne Meijerink-Bosman \email{m.l.meijerink@tilburguniversity.edu} [contributor]
  \item Diana Karimova \email{d.karimova@tilburguniversity.edu} [contributor]
  \item Mahdi Shafiee Kamalabad \email{m.shafiee@tilburguniversity.edu} [contributor]
  \item Roger Leenders \email{r.t.a.j.leenders@tilburguniversity.edu} [contributor]
  \item Joris Mulder \email{j.mulder3@tilburguniversity.edu} [contributor]
}

}
\keyword{internal}
