% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R, R/zzz.R
\docType{package}
\name{remstimate}
\alias{remstimate}
\title{remstimate}
\usage{
remstimate(
  reh = NULL,
  stats = NULL,
  method = c("MLE", "GD", "GDADAM", "BSIR", "HMC"),
  model = c("actor", "tie"),
  ncores = 1,
  prior = NULL,
  nsim = 1000,
  nchains = 2,
  burnin = 500,
  thin = 10,
  epochs = 200,
  learning_rate = 0.001,
  fast = FALSE,
  silent = TRUE,
  seed = sample(1:10000, nchains),
  ...
)
}
\arguments{
\item{reh}{a reh object (list of objects useful for the optimization)}

\item{stats}{a remstats object (array of statistics)}

\item{method}{optimization method to be used: MLE, GD, GDADAM, BSIR, HMC}

\item{model}{either actor or tie oriented}

\item{ncores}{number of threads for the parallelization (default = 1, no parallelization)}

\item{prior}{prior distribution when using Bayesian methods}

\item{nsim}{number of iterations in each chain}

\item{nchains}{number of chains to generate}

\item{burnin}{number of initial iterations to be added as burnin}

\item{thin}{number of steps to skip in the posterior draws of the HMC}

\item{epochs}{200 by defaut. It is the number of iteration used in the methods GD and GDADAM}

\item{learning_rate}{0.001 by default. It is the learning rate used in the methods GD and GDADAM}

\item{fast}{TRUE/FALSE whether to apply a faster computation of the likelihood}

\item{silent}{TRUE/FALSE if FALSE, progress of optimization status will be printed out}

\item{seed}{seed for reproducibility (yet to be integrated in the code)}
}
\value{
remstimate S3 object
}
\description{
A function that optimizes the likelihood of a Relational Event Model

The remstimate package provides a set of algorithms for the optimization of the relational even model likelihood
}
\section{\code{remstimate} functions}{

The remstimate functions ...
}

