% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R, R/zzz.R
\docType{package}
\name{remstimate}
\alias{remstimate}
\title{remstimate}
\usage{
remstimate(
  reh = NULL,
  stats = NULL,
  method = c("MLE", "GD", "GDADAM", "BSIR", "BPM", "HMC"),
  prior = NULL,
  proposal = NULL,
  n.chains = 2,
  n.iter = 2000,
  burnin = 1000,
  parallel = FALSE,
  nthreads = 1,
  seed = sample(1:10000, n.chains),
  silent = TRUE,
  ...
)
}
\arguments{
\item{reh}{a reh object (list of objects useful for the optimization)}

\item{stats}{a remstats object (array of statistics)}

\item{method}{optimization method to be used: trust, optim, GD, BSIR, HMC, BPM(?)}

\item{prior}{prior distribution when using Bayesian methods}

\item{proposal}{proposal distribution when using BSIR}

\item{n.chains}{number of chains to generate}

\item{n.iter}{number of iterations in each chain}

\item{burnin}{number of initial iterations to be added as burnin}

\item{parallel}{parallelize (especially with bayesian methods)}

\item{nthreads}{number of threads for the parallelization}

\item{seed}{seed for reproducibility}

\item{silent}{TRUE/FALSE if FALSE, progress of optimization status will be printed out}
}
\value{
remstimate S3 object
}
\description{
A function that returns maximum likelihood estimates of REM (interval timing only) by using the optim function

The remstimate package provides a set of algorithms for the optimization of the relational even model likelihood
}
\section{\code{remstimate} functions}{

The remstimate functions ...
}

