% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R, R/zzz.R
\docType{package}
\name{remstimate}
\alias{remstimate}
\title{remstimate}
\usage{
remstimate(
  pars_init = NULL,
  stats,
  event_binary,
  interevent_time,
  threads,
  method = c("Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", "Brent")
)
}
\arguments{
\item{stats}{an array of dimensions n_dyads*statistics*M (if stats is an array for relevent::rem then use aperm(a = stats, perm = c(2,3,1)) )}

\item{event_binary}{a matrix of dimensions M*n_dyads (for each time point it is a vector with 1 when the dyad is observed 0 for the others)}

\item{interevent_time}{a vector of interevent times}

\item{threads}{number of cores to use in the parallelization (i would change this parameter in 'n_threads')}
}
\value{
object list (the function saves also the output of optim)
}
\description{
A function that returns maximum likelihood estimates of REM (interval timing only) by using the optim function

The remstimate package provides a set of algorithms for the optimization of the relational even model likelihood
}
\section{\code{remstimate} functions}{

The remstimate functions ...
}

