% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R, R/zzz.R
\docType{package}
\name{remstimate}
\alias{remstimate}
\title{remstimate}
\usage{
remstimate(
  reh = NULL,
  stats = NULL,
  method = c("MLE", "GD", "GDADAM", "BSIR", "HMC"),
  model = c("actor", "tie"),
  ncores = 1,
  prior = NULL,
  nsim = 1000,
  bsir.samples = NULL,
  nchains = 2,
  burnin = 500,
  thin = 10,
  init = NULL,
  epochs = 200,
  learning_rate = 0.001,
  fast = FALSE,
  silent = TRUE,
  seed = sample(1:10000, nchains),
  ...
)
}
\arguments{
\item{reh}{an \code{reh} object (output object of the function \code{remify::reh})}

\item{stats}{a \code{remstats} object: when \code{model="tie"}, \code{stats} is an array of statictis with dimensions \code{[M x D x P]}: where \code{M} is the number of events, \code{D} is the number of possible dyads (full riskset), \code{P} is the number of statistics; if \code{model="actor"}, \code{stats} is a list of two arrays named \code{rate} and \code{choice} with dimensions \code{[M x N x P]}, where \code{N} are the actors (senders in the array \code{rate} receivers in the array \code{choice})}

\item{method}{optimization method to be used: \code{MLE}, \code{GD}, \code{GDADAM}, \code{BSIR}, \code{HMC}}

\item{model}{either \code{"actor"} or \code{"tie"} oriented (literature ref. here to DyNAM and to REM?)}

\item{ncores}{number of threads for the parallelization (\code{default = 1}, no parallelization)}

\item{prior}{prior distribution when using Bayesian methods}

\item{nsim}{number of iterations in each chain}

\item{bsir.samples}{number of samples from the proposal distribution when method is \code{"BSIR"}. When is left \code{NULL}, it is set to \code{nsim*3}}

\item{nchains}{number of chains to generate}

\item{burnin}{number of initial iterations to be added as burnin}

\item{thin}{number of steps to skip in the posterior draws of the HMC}

\item{init}{matrix of initial values for \code{"HMC"}}

\item{epochs}{200 by defaut. It is the number of iteration used in the methods \code{"GD"} and \code{"GDADAM"}}

\item{learning_rate}{0.001 by default. It is the learning rate used in the methods \code{"GD"} and \code{"GDADAM"}}

\item{fast}{\code{TRUE/FALSE} whether to apply a faster computation of the likelihood}

\item{silent}{\code{TRUE/FALSE} if \code{FALSE}, progress of optimization status will be printed out}

\item{seed}{seed for reproducibility (yet to be integrated in the code)}
}
\value{
remstimate S3 object
}
\description{
A function for the optimization of tie-oriented (Relational Event Model) or actor-oriented (DyNAM) likelihoods

The remstimate package provides a set of algorithms for the optimization of the relational even model likelihood
}
\section{\code{remstimate} functions}{

The remstimate functions ...
}

