% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R, R/zzz.R
\docType{package}
\name{remstimate}
\alias{remstimate}
\title{remstimate}
\usage{
remstimate(
  formula = NULL,
  data,
  approach = c("MLE", "Bayesian"),
  method = c("trust", "optim", "GD", "BSIR", "BPM", "HMC"),
  fast = FALSE,
  threshold = 0.5,
  stats = NULL,
  ...
)
}
\arguments{
\item{formula}{formula object (see 'remstats' for now)}

\item{data}{is a reh S3 object}

\item{approach}{estimation approach: MLE (Maximum Likelihood Estimates) or Bayesian}

\item{method}{optimization method to be used: trust/optim will use the corresponding functions, GD, BSIR, BPM, HMC}

\item{fast}{FALSE/TRUE whether to use the fast calculation of loglikelihood/gradient/hessian values inside the optimization algorithm}

\item{threshold}{percentage value indicating the minimum value for an actual improvment (in order to decide whether to use the fast approach for the likelihood or not).}

\item{stats}{an array of dimensions n_dyads*statistics*M (if stats is an array for relevent::rem then use aperm(a = stats, perm = c(2,3,1)) )}
}
\value{
object list (the function saves also the output of optim)
}
\description{
A function that returns maximum likelihood estimates of REM (interval timing only) by using the optim function

The remstimate package provides a set of algorithms for the optimization of the relational even model likelihood
}
\section{\code{remstimate} functions}{

The remstimate functions ...
}

