% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{residuals.remstimate}
\alias{residuals.remstimate}
\title{residuals.remstimate}
\usage{
\method{residuals}{remstimate}(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the 'residuals' method depending on the estimator used}
}
\description{
A function that extracts model residuals (Schoenfeld's residuals - add citation) from a 'remstimate' object.
}
\examples{

# No examples available at the moment

}
