% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remstimate.R
\name{waic}
\alias{waic}
\alias{waic.remstimate}
\title{waic}
\usage{
waic(object, ...)

\method{waic}{remstimate}(object, ...)
}
\arguments{
\item{object}{is a \code{remstimate} object}

\item{...}{further arguments to be passed to the 'waic' method}
}
\description{
A function that returns the WAIC (Watanabe-Akaike's Information Criterion) value in a 'remstimate' object
}
\section{Methods (by class)}{
\itemize{
\item \code{remstimate}: WAIC (Watanabe-Akaike's Information Criterion) value of a 'remstimate' object
}}

